separate header and sidebar
